Update and rename encs to encs.c
Fixed the name so the file actually has an extension. My bad. Also added support for arcade and holonomic controls, though code must be recompiled to switch. At the moment, I need to find the value for convertToRadians for the drive to use trig. Once that is done, I need to figure out how to account for strafing.